+++ core/src/datetime/mod.rs
@@ -9,3 +9,4 @@ mod util;
 pub use anchor::DateTimeAnchor;
 pub use loose::LooseDateTime;
 pub use util::RangePosition;
+pub(crate) use util::{STABLE_FORMAT_DATEONLY, STABLE_FORMAT_LOCAL};