+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - cli: Increase max todos
 - cli: Hide percent complete field in TUI todo editor when status is not InProcess (close #19)
 
+### FIXED
+
+- core: Query date only events
+
 ## [v0.6.0] - 2025-08-27
 
 ### Added